backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
httpx>=0.27,<1
python-multipart>=0.0.9,<1
websockets>=12,<17

# Ollama client (talks to local Ollama server)
ollama~=0.3.0

# Audio processing
numpy>=1.26,<3

# Speech-to-Text — Whisper (Hausa supported)
faster-whisper~=1.0.3

# Text-to-Speech — Piper (has ha_NG-openbible-medium Hausa voice)
piper-tts~=1.2.0

# Image & video generation — Diffusers + FLUX/Stable Diffusion / ModelScope T2V
diffusers>=0.30,<1
transformers>=4.44,<5
accelerate>=0.33,<1
torch~=2.3.0; python_version < "3.14"
torch>=2.9,<2.12; python_version >= "3.14"
Pillow>=10.4,<13

# Video export (used by diffusers.utils.export_to_video)
imageio>=2.34,<3
imageio-ffmpeg>=0.5,<1
